test(client): moved e2e to test/e2e directory
