+++ test/fixtures/reload-config/foo.js
@@ -1,3 +1,4 @@
 'use strict';
 
+// eslint-disable-next-line import/no-unresolved
 require('../../temp/main.css');
